requirements.txt
weaviate-client==3.0.0
sentence-transformers==3.1.1
torch==2.4.1
uvicorn[standard]==0.23.2
fastapi==0.103.1